Fix go vet warning about string conversion
string(int64) returns a single character string by treating the int64 as
a rune. That's not the behavior I expected when I wrote this function,
but it works well enough (I only care about avoiding collisions of
different block IDs), and it's almost certainly faster than doing an
Sprintf.

Casting to a rune explicitly suppresses the warning.